app/requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
bcrypt
psycopg2-binary
